requirements.txt
pandas~=1.3.4
pyspark~=3.2.0
pyarrow~=6.0.0

kafka-python~=2.0.2

yandexcloud==0.130.0
apache-airflow-providers-yandex==2.1.0